Cria superclasse abstrata Item para ItemSimples e ItemComposto
