+++ cooktop/src/main/java/br/edu/insper/desagil/aps6/cooktop/ItemSimples.java
@@ -1,26 +1,11 @@
 package br.edu.insper.desagil.aps6.cooktop;
 
-public class ItemSimples {
-	private String codigo;
-	private String nome;
+public class ItemSimples extends Item {
 	private int quantidade;
 
 	public ItemSimples(String codigo, String nome, int quantidade) {
-		this.codigo = codigo;
-		this.nome = nome;
+		super(codigo, nome);
 		this.quantidade = quantidade;
-	}
-
-	public String getCodigo() {
-		return codigo;
-	}
-
-	public String getNome() {
-		return nome;
-	}
-
-	public void setNome(String nome) {
-		this.nome = nome;
 	}
 
 	public int getQuantidade() {
